extract power level tags hook
